refactor: move BucketCapabilities and PathType to client module
