+++ client/src/main/java/com/google/cloud/gcs/analyticscore/client/BucketCapabilities.java
@@ -14,10 +14,20 @@
  * limitations under the License.
  */
 
-package com.google.cloud.gcs.analyticscore.common;
+package com.google.cloud.gcs.analyticscore.client;
 
-public enum PathType {
-  UNKNOWN,
-  FILE,
-  DIRECTORY
+/**
+ * Describes GCS bucket metadata capabilities, such as whether Hierarchical Namespace (HNS) is
+ * enabled. These capabilities influence query routing and optimization strategies.
+ */
+public class BucketCapabilities {
+  private final boolean hnsEnabled;
+
+  public BucketCapabilities(boolean hnsEnabled) {
+    this.hnsEnabled = hnsEnabled;
+  }
+
+  public boolean isHnsEnabled() {
+    return hnsEnabled;
+  }
 }
