+++ client/src/main/java/com/google/cloud/gcs/analyticscore/client/PathType.java
@@ -14,16 +14,11 @@
  * limitations under the License.
  */
 
-package com.google.cloud.gcs.analyticscore.common;
+package com.google.cloud.gcs.analyticscore.client;
 
-public class BucketCapabilities {
-  private final boolean hnsEnabled;
-
-  public BucketCapabilities(boolean hnsEnabled) {
-    this.hnsEnabled = hnsEnabled;
-  }
-
-  public boolean isHnsEnabled() {
-    return hnsEnabled;
-  }
+/** An optimization hint indicating the known type of a path to bypass redundant metadata checks. */
+public enum PathType {
+  UNKNOWN,
+  FILE,
+  DIRECTORY
 }
