+++ client/src/test/java/com/google/cloud/gcs/analyticscore/client/AnalyticsCacheManagerTest.java
@@ -19,7 +19,6 @@ package com.google.cloud.gcs.analyticscore.client;
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import com.google.cloud.gcs.analyticscore.common.BucketCapabilities;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.concurrent.atomic.AtomicInteger;
